Another attempt at problem 177
I proved that the angles have to be the same and it still doesn't work.
I loop through each possible rotation of the angles forwards and
backwards and it still doesn't work. This gosh darn problem has taken
up so much of my time and it still doesn't work. Gah. At this point I
should start randomly trying numbers around 750 000 000.